Added custom phone number editor and several front-end validators.
Currently, the client sends patches and posts to the server whether or not the front-end validation fails... not great. This isn't the biggest problem right now, so long as the server and client validation don't differ to much (at which point the user could see rejection messages while the server is in fact accepting the data).

I mostly just want to have SOME quick and fluid error messages pop up when obvious problems are present in the data being input into DataTable.