Add extra configurability to editors/formatters from Tables.js
Will slowly move more logic into the Tables and out of the generic. I've also removed broken tests and added some beginning tests for formatters - will hopefully move on to editors soon. The reality is that this front-end is a nightmare right now. It's difficult to change, takes for ever to reload and manually test. I need to be more strict with myself to leave the code better than I found it.